fix(delete-ip): Delete IP only if an IPID is present and IPPersistant is enabled
Delete IP only if an IPID is present and IPPersistant is enabled
Use of
the new methods for error handling

+++ scaleway/delete-ip.go
@@ -3,7 +3,6 @@ package scaleway
 import (
 	"github.com/rancher/machine/libmachine/log"
 	"github.com/scaleway/scaleway-sdk-go/api/instance/v1"
-	"github.com/scaleway/scaleway-sdk-go/scw"
 	"time"
 )
 
@@ -14,17 +13,24 @@ func (d *Driver) deleteIP() error {
 		return err
 	}
 
-	log.Infof("Deleting IP: %s", d.IPID)
+	if d.IPID != "" && d.IPPersistant {
+		log.Infof("Deleting IP: %s", d.IPID)
+		err = client.DeleteIP(&instance.DeleteIPRequest{
+			Zone: d.Zone,
+			IP:   d.IPID,
+		})
 
-	err = client.DeleteIP(&instance.DeleteIPRequest{
-		Zone: d.Zone,
-		IP:   d.IPID,
-	})
-
-	if err != nil && err.(*scw.ResponseError).StatusCode != 404 {
-		log.Infof("Delete of IP %s failed: %s, retrying in 10 seconds...", d.IPID, err.Error())
-		time.Sleep(10 * time.Second)
-		return d.deleteIP()
+		if err != nil {
+			if IsScwError(err) {
+				if GetErrorStatus(err) != 404 {
+					log.Infof("Delete of IP %s failed: %s, retrying in 10 seconds...", d.IPID, err.Error())
+					time.Sleep(10 * time.Second)
+					return d.deleteIP()
+				}
+			}
+			log.Errorf("Delete of IP %s failed: %s", d.IPID, err.Error())
+			return err
+		}
 	}
 
 	return nil
